requirements.txt
Flask==2.2.2
Flask_Admin==1.6.1
Flask_Mail==0.9.1
Flask_MySQLdb==1.0.1
flask_sqlalchemy==3.0.5
folium==0.14.0
mysql_connector_repackaged==0.3.1
pandas==1.5.2
plotly==5.12.0
PyMySQL==1.0.3
PyYAML==6.0
PyYAML==6.0.1
